01dl_uisbp_lte/requirements_tf2.0.txt
numpy
opencv_python==3.4.0.12
Keras==2.1.6
labelme==3.16.7
scipy==1.3.1
setuptools==41.0.1
matplotlib==3.1.1
tqdm==4.35.0
tensorflow-gpu==2.0.0rc1
Pillow
pydicom
ruamel.base==1.0.0
seaborn==0.9.0
scikit_learn==0.21.3
typing==3.7.4.1
